pyproject.toml
[project]
name = "open-terminal"
version = "0.11.26"
description = "A remote terminal API."
readme = "README.md"
authors = [
    { name = "Timothy Jaeryang Baek", email = "tim@openwebui.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.22",
    "aiofiles>=25.1.0",
    "pypdf>=5.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.0",
    "striprtf>=0.0.26",
    "xlrd>=2.0.0",
    "nbclient>=0.10.0",
    "ipykernel>=6.0.0",
    "pywinpty>=2.0.0; sys_platform == 'win32'",
]

[project.optional-dependencies]
mcp = ["fastmcp>=2.0.0"]

[project.scripts]
open-terminal = "open_terminal.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["open_terminal"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
